Aplicação das seguintes alterações:
- Na classe "Debito" - atualização do tipo de dado do campo "DebitoId" de int para long
- Na classe "DebitosController" foram adicionados os métodos de pesquisa de débitos ativos e de efetuar o pagamento do mesmo.

+++ src/MvcBiblioteca/Biblioteca.Dominio/Debito.cs
@@ -8,7 +8,7 @@ namespace Biblioteca.Dominio
 {
     public class Debito
     {
-        public int DebitoId { get; set; }
+        public long DebitoId { get; set; }
 
         [DisplayName("Usuário")]
         public Usuario UsuarioDeb { get; set; }
